docker files created

+++ API/Controllers/lost_item_controller.py
@@ -73,7 +73,7 @@ def update_lost_item(db: Session, lost_item_id: int, lost_item_schemas: lost_ite
         return _u
     raise HTTPException(status_code=404, detail="Not Found")
 
-def delete_lost_item(db: Session, lost_item_id: float):
+def delete_lost_item(db: Session, lost_item_id: int):
     delete = db.query(models.Lost_Item).filter(models.Lost_Item.lost_item_id == lost_item_id).first()
     if delete is None:
         raise HTTPException(status_code=404, detail="Not Found")
